Add image to README.md
The image shows an example of what the terminal looks like when in an SSH session, in a virtual environment, in a github repository, and in a directory whose path is long. It also shows examples of the terminal output when running Ctrl+C, Ctrl+V, and a command with an exit status of 1. I've never worked with images in a README.md before so I'm not sure this'll work.